% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-snapshot.R
\name{app_expect_download}
\alias{app_expect_download}
\title{Expect a downloadable file}
\usage{
app_expect_download(app, id, ..., name = NULL, cran = FALSE)
}
\arguments{
\item{app}{A \link{ShinyDriver2} object.}

\item{id}{Output id of \code{\link[shiny:downloadButton]{shiny::downloadButton()}}/\code{\link[shiny:downloadButton]{shiny::downloadLink()}}}

\item{...}{Must be empty. Allows for parameter expansion.}

\item{name}{File name to save file to. The default, \code{NULL},
generates an ascending sequence of names: \verb{001.download},
\verb{002.download}, etc.}

\item{cran}{Should these expectations be verified on CRAN? By default,
they are not, because snapshot tests tend to be fragile because they
often rely on minor details of dependencies.}
}
\description{
Expect a downloadable file
}
