% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-expect-other.R
\name{app_expect_text}
\alias{app_expect_text}
\alias{app_expect_html}
\title{Expect App DOM or text snapshot}
\usage{
app_expect_text(app, selector, ..., cran = FALSE)

app_expect_html(app, selector, ..., outer_html = FALSE, cran = FALSE)
}
\arguments{
\item{app}{A \link{ShinyDriver2} object.}

\item{selector}{A DOM selector to be passed into jQuery}

\item{...}{Must be empty. Allows for parameter expansion.}

\item{cran}{Should these expectations be verified on CRAN? By default,
they are not, because snapshot tests tend to be fragile because they
often rely on minor details of dependencies.}

\item{outer_html}{If \code{TRUE}, the full DOM structure will be returned (\code{tag.outerHTML}).
If \code{FALSE}, the full DOM structure of the child elements will be returned (\code{tag.innerHTML}).}
}
\description{
These methods will find all matching elements and extract their respective values.
These values will then be stored as a snapshot.
The snapshot \code{variant} will be extracted from the \code{app} object directly.
}
\section{Functions}{
\itemize{
\item \code{app_expect_text}: This method will extract the text value of all matching elements via \code{tag.textContent}.
This method is more robust to internal package change as only the text values will be return.
When possible, use \code{app_expect_text()} over \code{app_expect_html()} to allow package authors room to change.

\item \code{app_expect_html}: This method will extract the relevant DOM structure of all matching elements.
This method is great for testing the full DOM structure of particular HTML elements.
It is recommended to only use this method on DOM elements that you have full control over.
This will help avoid false-positives when underlying packages may update.
}}

