% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote_methods.R
\name{chromote_eval_script}
\alias{chromote_eval_script}
\title{Evaluate a script on the window context}
\usage{
chromote_eval_script(
  chromote,
  js_script,
  awaitPromise = TRUE,
  arguments = NULL,
  timeout = 10 * 1000,
  ...
)
}
\arguments{
\item{arguments}{V}
}
\description{
Evaluate a script on the window context
}
