% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote_methods.R
\name{chromote_eval_script_callback}
\alias{chromote_eval_script_callback}
\title{Evaluate on the window context, blocking until the wrapping promise is resolved}
\usage{
chromote_eval_script_callback(
  chromote,
  js_script,
  awaitPromise = TRUE,
  arguments = NULL,
  timeout = 10 * 1000,
  ...
)
}
\description{
Evaluate on the window context, blocking until the wrapping promise is resolved
}
