% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote_methods.R
\name{chromote_wait_for_condition}
\alias{chromote_wait_for_condition}
\title{Have chromote wait for a js condition to be met}
\usage{
chromote_wait_for_condition(
  chromote,
  condition_js,
  ...,
  timeout = 15 * 1000,
  interval = 100,
  wait_ = TRUE
)
}
\arguments{
\item{condition_js}{A piece of JavaScript code that should eventually evaluate to a \href{https://developer.mozilla.org/en-US/docs/Glossary/Truthy}{\code{true}thy value}.}

\item{interval}{How long (milliseconds) Chrome should wait between checking \code{condition_js}}
}
\value{
\code{TRUE} if expression evaluates to \code{TRUE} without error, before
timeout. Otherwise returns \code{FALSE}.
}
\description{
Have chromote wait for a js condition to be met
}
\section{Functions}{
\itemize{
\item \code{chromote_wait_for_condition}: Waits for \code{condition_js} to return a \code{true}thy value. In this con
}}

