% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote_methods.R
\name{chromote_wait_for_condition}
\alias{chromote_wait_for_condition}
\title{Have chromote wait for a js condition to be met}
\usage{
chromote_wait_for_condition(
  chromote,
  condition_js,
  ...,
  timeout = 15 * 1000,
  interval = 100,
  wait_ = TRUE
)
}
\value{
\code{TRUE} if expression evaluates to \code{TRUE} without error, before
timeout. Otherwise returns \code{FALSE}.
}
\description{
Have chromote wait for a js condition to be met
}
