% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-file-screenshot.R
\name{compare_file_screenshot}
\alias{compare_file_screenshot}
\title{Compare screenshots given tolerance}
\usage{
compare_file_screenshot(old, new, tolerance = 10)
}
\arguments{
\item{old}{Current screenshot file}

\item{new}{New screenshot file}

\item{tolerance}{How many units of difference to allow}
}
\description{
\pkg{chromote} produces screenshot images with inconsistent rounding for areas like
the corner of a text area. This function compares two images and allows for a \code{tolerance}
of so many units in each RGBA color channel.
}
\details{
It is suggested to use this method with \code{\link{AppDriver}}\verb{$expect_screenshot()} to compare
screenshots given a particular tolerance.
}
