% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-screenshot-threshold.R
\name{compare_screenshot_threshold}
\alias{compare_screenshot_threshold}
\title{Compare screenshots given threshold value}
\usage{
compare_screenshot_threshold(old, new, ..., threshold = NULL, kernel_size = 5)
}
\arguments{
\item{old}{Current screenshot file}

\item{new}{New screenshot file}

\item{...}{Must be empty. Allows for parameter expansion.}

\item{threshold}{How many units of difference to allow}

\item{kernel_size}{How many pixels tall and wide should the convolution be applied to}
}
\description{
\pkg{chromote} produces screenshot images with inconsistent rounding for areas like
the corner of a text area. This function compares two images and allows for a \code{threshold}
of so many units in each RGBA color channel.
}
\details{
It is suggested to use this method with \code{\link{AppDriver}}\verb{$expect_screenshot()} to compare
screenshots given a particular threshold value.
}
