% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-appshot.R
\name{expect_snapshot_app}
\alias{expect_snapshot_app}
\title{Expect a shinytest2 snapshot}
\usage{
expect_snapshot_app(
  app,
  ...,
  name = NULL,
  items = NULL,
  screenshot = NULL,
  cran = FALSE
)
}
\arguments{
\item{app}{A \link{ShinyDriver2} object.}

\item{...}{Must be empty. Allows for parameter expansion.}

\item{name}{The prefix name to be used for the snapshot. By default, this uses the name supplied to \code{app} on initialization.}

\item{items}{Elements to only be included in the snapshot. If supplied, can contain \code{inputs}, \code{output}, and \code{export}. Each value of \code{items} can either be \code{TRUE} (for all values) or a character list of names to use.}

\item{screenshot}{A boolean indicating whether to take a screenshot.}

\item{cran}{Should these expectations be verified on CRAN? By default,
they are not, because snapshot tests tend to be fragile because they
often rely on minor details of dependencies.}
}
\description{
Expect a shinytest2 snapshot
}
