% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-expect.R
\name{expect_snapshot_app_js}
\alias{expect_snapshot_app_js}
\title{Expect snapshot of JS script output}
\usage{
expect_snapshot_app_js(
  app,
  script,
  ...,
  arguments = list2(),
  post_script = NULL,
  cran = FALSE
)
}
\arguments{
\item{app}{A \link{ShinyDriver2} object.}

\item{script}{A string containing the JS script to be executed.}

\item{...}{Must be empty. Allows for parameter expansion.}

\item{arguments}{A list of arguments to be passed to the script.}

\item{post_script}{A function to be called on the result of the script before taking the snapshot.
\code{\link[=expect_snapshot_app_dom]{expect_snapshot_app_dom()}} and \code{\link[=expect_snapshot_app_text]{expect_snapshot_app_text()}} both use \code{\link[=unlist]{unlist()}}.}

\item{cran}{Should these expectations be verified on CRAN? By default,
they are not, because snapshot tests tend to be fragile because they
often rely on minor details of dependencies.}
}
\description{
This is a building block function that should be called by other functions.
For example, \code{\link[=expect_snapshot_app_dom]{expect_snapshot_app_dom()}} and \code{\link[=expect_snapshot_app_text]{expect_snapshot_app_text()}} are thin wrappers around this function.
}
