% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{expect_st2_snapshot}
\alias{expect_st2_snapshot}
\title{Test Shinytest2 code chunk}
\usage{
expect_st2_snapshot(
  app,
  ...,
  name = NULL,
  items = NULL,
  screenshot = NULL,
  cran = FALSE,
  transform = NULL,
  variant = os_name_and_r_version()
)
}
\description{
Used to shim in extra information (platform and R version) into the \code{desc}ription of the test.
}
\details{
Since \code{shinytest::test_that} will know which platform is being run, other platform's test snaps will not be auto-deleted due to not being used.
}
