% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{os_name}
\alias{os_name}
\alias{r_version}
\alias{os_name_and_r_version}
\title{Get the name of the OS}
\usage{
os_name()

r_version()

os_name_and_r_version()
}
\description{
Returns the name of the current OS. This can be useful for the \code{variant} when
running \code{\link[testthat:test_dir]{testthat::test_dir()}}.
}
\section{Functions}{
\itemize{
\item \code{os_name}: Operating system name

\item \code{r_version}: Return the Major and Minor version of the OS. No Patch version

\item \code{os_name_and_r_version}: Return the \code{os_name()} and \code{r_version()} join by a \code{-}
}}

