% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_variant}
\alias{platform_variant}
\title{Platform specific variant}
\usage{
platform_variant(..., os_name = TRUE, r_version = TRUE)
}
\arguments{
\item{...}{Must be empty. Allows for parameter expansion.}

\item{os_name}{if \code{TRUE}, include the OS name in the output}

\item{r_version}{if \code{TRUE}, include the major and minor version of the R version, no patch version}
}
\description{
Returns a string to be used within \pkg{testthat}'s' snapshot testing. Currently, the Operating System
and R Version (major, minor, no patch version) are returned.
}
\details{
If more information is needed in the future to distinguish standard testing environments, this function
will be updated accordingly.
}
\seealso{
\code{\link[testthat:test_dir]{testthat::test_dir()}}
}
