% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record-test.R
\name{register_input_processor}
\alias{register_input_processor}
\alias{get_input_processors}
\title{Register an input processor for the test recorder}
\usage{
register_input_processor(input_type, processor)

get_input_processors()
}
\arguments{
\item{input_type}{The name of an input type, for example,
\code{"mypkg.numberinput"}.}

\item{processor}{An input processor function.}
}
\description{
\code{register_input_processor()} registers an input processor which will be used by
the test recorder. The input processor function should take one parameter,
\code{value}, and return a string of R code which returns the desired value.

\code{get_input_processors()} returns a named list of all registered input processor
functions.
}
\keyword{internal}
