% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-app.R
\name{test_app}
\alias{test_app}
\title{Test shiny application with testthat}
\usage{
test_app(app_dir = "../", env = shiny::loadSupport(app_dir), ...)
}
\arguments{
\item{app_dir}{The base directory for the Shiny application}

\item{env}{Use the shiny application's environment after sourcing the R folder}

\item{...}{Parameters passed to \code{\link[testthat:test_dir]{testthat::test_dir()}}}
}
\description{
To be used within \code{./tests/testthat.R}.
}
\details{
Example usage:\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Interactive usage
path_to_app <- "."
shinytest2::test_app(path_to_app)

# ./tests/testthat.R
shinytest2::test_app()
}\if{html}{\out{</div>}}
}
