% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_shinytest2}
\alias{use_shinytest2}
\title{Use \pkg{shinytest2} methods}
\usage{
use_shinytest2(
  app_dir = ".",
  ...,
  open = rlang::is_interactive(),
  quiet = FALSE,
  overwrite = FALSE,
  actions = c("runner", "test_file", "ignore", "package")
)
}
\arguments{
\item{app_dir}{The base directory for the Shiny application}

\item{...}{Must be empty. Allows for parameter expansion.}

\item{open}{If \code{TRUE}, the test file will be opened in an editor via
\code{\link[=file.edit]{file.edit()}} after saving.}

\item{quiet}{If \code{TRUE}, console output will be suppressed.}

\item{overwrite}{If \code{TRUE}, the test file or test runner will be overwritten.}

\item{actions}{Actions that can be performed:
\itemize{
\item \code{"runner"}: Create a \pkg{shinytest2} test runner at \code{./tests/testthat.R}
\item \code{"test_file"}: Create a \pkg{shinytest2} test file at \code{./tests/testthat/test-shinytest2.R}
\item \code{"ignore"}: Add entries to \code{.Rbuildignore} and \code{.gitignore} to ignore new debug screenshots. (\verb{*_.new.png})
\item \code{"package"}: Add \pkg{shinytest2} to \code{Suggests} in the \code{DESCRIPTION} file.
}}
}
\description{
This unified method initializes many different useful features when using
\pkg{shinytest2}. See \code{actions} below for more details.
}
\examples{
\dontrun{use_shinytest2()}
}
