% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-widget.R
\name{Widget2}
\alias{Widget2}
\title{A Shiny Widget}
\description{
A \code{Widget2} object represents a Shiny input or output control, and provides
methods for finer grained interaction.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_chromote_session}{\code{Widget2$get_chromote_session()}}
\item \href{#method-new}{\code{Widget2$new()}}
\item \href{#method-get_type}{\code{Widget2$get_type()}}
\item \href{#method-get_iotype}{\code{Widget2$get_iotype()}}
\item \href{#method-get_value}{\code{Widget2$get_value()}}
\item \href{#method-set_value}{\code{Widget2$set_value()}}
\item \href{#method-click}{\code{Widget2$click()}}
\item \href{#method-upload_file}{\code{Widget2$upload_file()}}
\item \href{#method-clone}{\code{Widget2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_chromote_session"></a>}}
\if{latex}{\out{\hypertarget{method-get_chromote_session}{}}}
\subsection{Method \code{get_chromote_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$get_chromote_session()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new \code{Widget2}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$new(
  name,
  node_id,
  type,
  iotype = c("input", "output"),
  chromote_session
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a Shiny widget.}

\item{\code{type}}{Widget2 type}

\item{\code{iotype}}{Input/output type.}

\item{\code{nodeId}}{Node Id of an Element from the \code{chromote} session.}

\item{\code{chromote}}{Chromote Session}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-get_type}{}}}
\subsection{Method \code{get_type()}}{
Widget2 type, e.g. \code{textInput}, \code{selectInput}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$get_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_iotype"></a>}}
\if{latex}{\out{\hypertarget{method-get_iotype}{}}}
\subsection{Method \code{get_iotype()}}{
Is this an input or output control?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$get_iotype()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-get_value}{}}}
\subsection{Method \code{get_value()}}{
Get current value of control.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$get_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-set_value}{}}}
\subsection{Method \code{set_value()}}{
Set value of control.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$set_value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value to set for the widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-click"></a>}}
\if{latex}{\out{\hypertarget{method-click}{}}}
\subsection{Method \code{click()}}{
Sends a jQuery click to the element.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$click()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_file"></a>}}
\if{latex}{\out{\hypertarget{method-upload_file}{}}}
\subsection{Method \code{upload_file()}}{
Upload a file to a \code{\link[shiny:fileInput]{shiny::fileInput()}}.
It fails for other types of widgets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$upload_file(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Path to file to upload}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
