% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-app.R
\name{local_app_support}
\alias{local_app_support}
\alias{with_app_support}
\title{Attach the Shiny application's support environment}
\usage{
local_app_support(app_dir = "../../", envir = rlang::caller_env())

with_app_support(expr, app_dir = "../../", envir = rlang::caller_env())
}
\arguments{
\item{app_dir}{The base directory for the Shiny application.}

\item{envir}{The environment in which the files in the \verb{R/} directory should
be evaluated.}

\item{expr}{An expression to evaluate within the support environment.}
}
\description{
Executes all \code{./R} files and \code{global.R} into a temp environment that is
attached appropriately. This is useful when wanting access to functions or
values created in the \code{./R} folder for testing purposes.
}
\details{
Loading these files is not automatically performed by \code{test_app()} and must
be called within your unit tests to access support file objects.
}
\section{Functions}{
\itemize{
\item \code{local_app_support()}: Temporarily attach the Shiny application's support
environment into the current environment.

\item \code{with_app_support()}: For the provided \code{expr}, attach the Shiny
application's support environment into the current environment.

}}
\examples{
\dontrun{
# ./R/utils.R
n <- 42

#' # ./tests/testthat/test-utils.R
test_that("Can access support environment", {
  shinytest2::local_app_support()
  expect_equal(n, 42)
})
}
}
