% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote_methods.R
\name{chromote_execute_script}
\alias{chromote_execute_script}
\alias{chromote_execute_script_callback}
\title{Execute a JavaScript script}
\usage{
chromote_execute_script(
  chromote,
  js_script,
  ...,
  awaitPromise = wait_,
  arguments = list(),
  timeout = 10 * 1000,
  wait_ = TRUE
)

chromote_execute_script_callback(
  chromote,
  js_script,
  ...,
  arguments = list(),
  timeout = 15 * 1000,
  wait_ = TRUE
)
}
\arguments{
\item{chromote}{A ChromoteSession object}

\item{js_script}{A string containing the script to be evaluated}

\item{...}{Arguments passed to \code{chromote$Runtime$evaluate}, such as \code{wait_}}

\item{awaitPromise}{If \code{TRUE} (default), the function will wait for the \code{js_script} promise to be resolved. If \code{wait_ == FALSE}, \code{awaitPromise} will be set to \code{FALSE}.}

\item{arguments}{An unnamed list of arguments to be passed into the \code{js_script}}

\item{timeout}{The maximum time (milliseconds) \code{chromote} will wait for the \code{js_script} promise to be resolved}
}
\description{
Given \code{wait_ = TRUE}, \code{chromote_execute_script()} will block the testing R session
until the script has completed the JS execution tick within the headless browser.
}
\section{Functions}{
\itemize{
\item \code{chromote_execute_script}: Executes the supplied Javascript script (\code{js_script}) within a function. The function has the \code{window} context and access to \code{arguments} supplied.

\item \code{chromote_execute_script_callback}: Executes the supplied Javascript script (\code{js_script}) within a function. The function has the \code{window} context and access to \code{arguments} supplied. An extra argument (\code{resolve(val)}) is added to the \code{arguments} list. If \code{wait_ = TRUE}, then \code{chromote_execute_script_callback()} will block the main R session until \code{resolve()} has been called.
}}

