% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{record_test}
\alias{record_test}
\title{Launch test event recorder for a Shiny app}
\usage{
record_test(
  app = ".",
  ...,
  seed = NULL,
  load_timeout = 10000,
  debug = "shiny_console",
  shiny_args = list()
)
}
\arguments{
\item{app}{A \link{AppDriver} object, or path to a Shiny
application.}

\item{...}{Used for parameter expansion}

\item{seed}{A random seed to set before running the app. This seed will also
be used in the test script.}

\item{load_timeout}{Maximum time to wait for the Shiny application to load, in
milliseconds. If a value is provided, it will be saved in the test script.}

\item{debug}{start the underlying \link{AppDriver} in \code{debug}
mode and print those debug logs to the R console once recording is
finished. The default, \code{'shiny_console'}, captures and prints R
console output from the recorded R shiny process. Any value that the
\code{debug} argument in \link{AppDriver} accepts may be used
(e.g., \code{'none'} may be used to completely suppress the driver logs).}

\item{shiny_args}{A list of options to pass to \code{runApp()}. If a value
is provided, it will be saved in the test script.}
}
\description{
Launch test event recorder for a Shiny app
}
