% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{record_test}
\alias{record_test}
\title{Launch test event recorder for a Shiny app}
\usage{
record_test(
  app = ".",
  ...,
  seed = NULL,
  load_timeout = 10 * 1000,
  shiny_args = list()
)
}
\arguments{
\item{app}{A \link{AppDriver} object, or path to a Shiny
application.}

\item{...}{Must be empty. Allows for parameter expansion.}

\item{seed}{A random seed to set before running the app. This seed will also
be used in the test script.}

\item{load_timeout}{Maximum time to wait for the Shiny application to load, in
milliseconds. If a value is provided, it will be saved in the test script.}

\item{shiny_args}{A list of options to pass to \code{runApp()}. If a value
is provided, it will be saved in the test script.}
}
\description{
Launch test event recorder for a Shiny app
}
