% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-driver.R
\name{ShinyDriver2}
\alias{ShinyDriver2}
\title{Remote control a Shiny app running in a headless browser}
\description{
This class starts a Shiny app in a new R session, along with a \pkg{chromote}
headless browser that can be used to simulate user actions. This provides
a full simulation of a Shiny app so that you can test user interactions
with a live app.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-expect_appshot}{\code{ShinyDriver2$expect_appshot()}}
\item \href{#method-get_chromote_session}{\code{ShinyDriver2$get_chromote_session()}}
\item \href{#method-get_debug_log}{\code{ShinyDriver2$get_debug_log()}}
\item \href{#method-enable_debug_log_messages}{\code{ShinyDriver2$enable_debug_log_messages()}}
\item \href{#method-logEvent}{\code{ShinyDriver2$logEvent()}}
\item \href{#method-getEventLog}{\code{ShinyDriver2$getEventLog()}}
\item \href{#method-executeScript}{\code{ShinyDriver2$executeScript()}}
\item \href{#method-executeScriptCallback}{\code{ShinyDriver2$executeScriptCallback()}}
\item \href{#method-expectSnapshotJS}{\code{ShinyDriver2$expectSnapshotJS()}}
\item \href{#method-getAllValues}{\code{ShinyDriver2$getAllValues()}}
\item \href{#method-new}{\code{ShinyDriver2$new()}}
\item \href{#method-setInputs}{\code{ShinyDriver2$setInputs()}}
\item \href{#method-uploadFile}{\code{ShinyDriver2$uploadFile()}}
\item \href{#method-get_path}{\code{ShinyDriver2$get_path()}}
\item \href{#method-takeScreenshot}{\code{ShinyDriver2$takeScreenshot()}}
\item \href{#method-expectDownload}{\code{ShinyDriver2$expectDownload()}}
\item \href{#method-stop}{\code{ShinyDriver2$stop()}}
\item \href{#method-waitFor}{\code{ShinyDriver2$waitFor()}}
\item \href{#method-waitForShiny}{\code{ShinyDriver2$waitForShiny()}}
\item \href{#method-waitForValue}{\code{ShinyDriver2$waitForValue()}}
\item \href{#method-findWidget}{\code{ShinyDriver2$findWidget()}}
\item \href{#method-getValue}{\code{ShinyDriver2$getValue()}}
\item \href{#method-setValue}{\code{ShinyDriver2$setValue()}}
\item \href{#method-click}{\code{ShinyDriver2$click()}}
\item \href{#method-listWidgets}{\code{ShinyDriver2$listWidgets()}}
\item \href{#method-checkUniqueWidgetNames}{\code{ShinyDriver2$checkUniqueWidgetNames()}}
\item \href{#method-setWindowSize}{\code{ShinyDriver2$setWindowSize()}}
\item \href{#method-getWindowSize}{\code{ShinyDriver2$getWindowSize()}}
\item \href{#method-clone}{\code{ShinyDriver2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expect_appshot"></a>}}
\if{latex}{\out{\hypertarget{method-expect_appshot}{}}}
\subsection{Method \code{expect_appshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$expect_appshot(
  ...,
  name = NULL,
  items = NULL,
  screenshot = NULL,
  error = FALSE,
  cran = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a shiny widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_chromote_session"></a>}}
\if{latex}{\out{\hypertarget{method-get_chromote_session}{}}}
\subsection{Method \code{get_chromote_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$get_chromote_session()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_debug_log"></a>}}
\if{latex}{\out{\hypertarget{method-get_debug_log}{}}}
\subsection{Method \code{get_debug_log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$get_debug_log(type = c("all", debug_log_types()))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-enable_debug_log_messages"></a>}}
\if{latex}{\out{\hypertarget{method-enable_debug_log_messages}{}}}
\subsection{Method \code{enable_debug_log_messages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$enable_debug_log_messages(enable = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logEvent"></a>}}
\if{latex}{\out{\hypertarget{method-logEvent}{}}}
\subsection{Method \code{logEvent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$logEvent(event, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEventLog"></a>}}
\if{latex}{\out{\hypertarget{method-getEventLog}{}}}
\subsection{Method \code{getEventLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$getEventLog()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-executeScript"></a>}}
\if{latex}{\out{\hypertarget{method-executeScript}{}}}
\subsection{Method \code{executeScript()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$executeScript(script, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-executeScriptCallback"></a>}}
\if{latex}{\out{\hypertarget{method-executeScriptCallback}{}}}
\subsection{Method \code{executeScriptCallback()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$executeScriptCallback(script, ..., timeout_ = 15 * 1000)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expectSnapshotJS"></a>}}
\if{latex}{\out{\hypertarget{method-expectSnapshotJS}{}}}
\subsection{Method \code{expectSnapshotJS()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$expectSnapshotJS(script, ..., post_script = NULL, cran = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAllValues"></a>}}
\if{latex}{\out{\hypertarget{method-getAllValues}{}}}
\subsection{Method \code{getAllValues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$getAllValues(input = TRUE, output = TRUE, export = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$new(
  path = ".",
  ...,
  load_timeout = NULL,
  screenshot = TRUE,
  check_names = TRUE,
  name = NULL,
  variant = getOption("shinytest2.variant", os_name_and_r_version()),
  debug = c("none", "all", debug_log_types()),
  seed = NULL,
  clean_logs = TRUE,
  shiny_args = list(),
  render_args = NULL,
  options = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a shiny widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setInputs"></a>}}
\if{latex}{\out{\hypertarget{method-setInputs}{}}}
\subsection{Method \code{setInputs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$setInputs(
  ...,
  wait_ = TRUE,
  values_ = TRUE,
  timeout_ = 3000,
  allow_input_no_binding_ = FALSE,
  priority_ = c("input", "event")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wait_}}{Wait until all reactive updates have completed?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-uploadFile"></a>}}
\if{latex}{\out{\hypertarget{method-uploadFile}{}}}
\subsection{Method \code{uploadFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$uploadFile(..., wait_ = TRUE, values_ = TRUE, timeout_ = 3000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wait_}}{Wait until all reactive updates have completed?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_path"></a>}}
\if{latex}{\out{\hypertarget{method-get_path}{}}}
\subsection{Method \code{get_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$get_path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-takeScreenshot"></a>}}
\if{latex}{\out{\hypertarget{method-takeScreenshot}{}}}
\subsection{Method \code{takeScreenshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$takeScreenshot(
  filename = NULL,
  ...,
  screenshot_args = list(),
  delay = 0,
  selector = "html",
  wait_ = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wait_}}{Wait until all reactive updates have completed?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expectDownload"></a>}}
\if{latex}{\out{\hypertarget{method-expectDownload}{}}}
\subsection{Method \code{expectDownload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$expectDownload(id, ..., name = NULL, cran = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a shiny widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waitFor"></a>}}
\if{latex}{\out{\hypertarget{method-waitFor}{}}}
\subsection{Method \code{waitFor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$waitFor(expr, timeout = 3 * 1000, interval = 100)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waitForShiny"></a>}}
\if{latex}{\out{\hypertarget{method-waitForShiny}{}}}
\subsection{Method \code{waitForShiny()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$waitForShiny(timeout = 3 * 1000, interval = 100)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waitForValue"></a>}}
\if{latex}{\out{\hypertarget{method-waitForValue}{}}}
\subsection{Method \code{waitForValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$waitForValue(
  name,
  ignore = list(NULL, ""),
  iotype = c("input", "output", "export"),
  timeout = 10000,
  check_interval = 400
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a shiny widget.}

\item{\code{iotype}}{Type of the Shiny widget. Usually shinytest2 finds the widgets
by their name, so this is only needed if you use the same name for an
input and output widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findWidget"></a>}}
\if{latex}{\out{\hypertarget{method-findWidget}{}}}
\subsection{Method \code{findWidget()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$findWidget(name, iotype = c("auto", "input", "output"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a shiny widget.}

\item{\code{iotype}}{Type of the Shiny widget. Usually shinytest2 finds the widgets
by their name, so this is only needed if you use the same name for an
input and output widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getValue"></a>}}
\if{latex}{\out{\hypertarget{method-getValue}{}}}
\subsection{Method \code{getValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$getValue(name, iotype = c("auto", "input", "output"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a shiny widget.}

\item{\code{iotype}}{Type of the Shiny widget. Usually shinytest2 finds the widgets
by their name, so this is only needed if you use the same name for an
input and output widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setValue"></a>}}
\if{latex}{\out{\hypertarget{method-setValue}{}}}
\subsection{Method \code{setValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$setValue(name, value, iotype = c("auto", "input", "output"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a shiny widget.}

\item{\code{iotype}}{Type of the Shiny widget. Usually shinytest2 finds the widgets
by their name, so this is only needed if you use the same name for an
input and output widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-click"></a>}}
\if{latex}{\out{\hypertarget{method-click}{}}}
\subsection{Method \code{click()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$click(name, iotype = c("auto", "input", "output"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a shiny widget.}

\item{\code{iotype}}{Type of the Shiny widget. Usually shinytest2 finds the widgets
by their name, so this is only needed if you use the same name for an
input and output widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listWidgets"></a>}}
\if{latex}{\out{\hypertarget{method-listWidgets}{}}}
\subsection{Method \code{listWidgets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$listWidgets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkUniqueWidgetNames"></a>}}
\if{latex}{\out{\hypertarget{method-checkUniqueWidgetNames}{}}}
\subsection{Method \code{checkUniqueWidgetNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$checkUniqueWidgetNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setWindowSize"></a>}}
\if{latex}{\out{\hypertarget{method-setWindowSize}{}}}
\subsection{Method \code{setWindowSize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$setWindowSize(width, height)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWindowSize"></a>}}
\if{latex}{\out{\hypertarget{method-getWindowSize}{}}}
\subsection{Method \code{getWindowSize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$getWindowSize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyDriver2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
