% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-app.R
\name{test_app}
\alias{test_app}
\title{Run tests for a Shiny application}
\usage{
test_app(app_dir = ".", ..., reporter = missing_arg(), package = missing_arg())
}
\arguments{
\item{...}{Ignored. Used for parameter expansion}

\item{reporter}{Reporter to use to summarise output. Can be supplied
as a string (e.g. "summary") or as an R6 object
(e.g. \code{SummaryReporter$new()}).

See \link[testthat]{Reporter} for more details and a list of built-in reporters.}

\item{package}{If these tests belong to a package, the name of the package.}

\item{appDir}{Path to directory containing a Shiny app (e.g. \code{app.R}) or
single interactive \code{.Rmd}.}
}
\description{
Run tests for a Shiny application
}
\seealso{
\code{\link[testthat:snapshot_accept]{testthat::snapshot_review()}} and \code{\link[testthat:snapshot_accept]{testthat::snapshot_accept()}} if
you want to compare or update snapshots after testing. In most cases, the
user is prompted to do these tasks interactively, but there are also times
where it is useful to call these functions from the console.
}
