% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-widget.R
\name{Widget2}
\alias{Widget2}
\title{A Shiny Widget}
\description{
A \code{Widget2} object represents a Shiny input or output control, and provides
methods for finer grained interaction.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Widget2$new()}}
\item \href{#method-getName}{\code{Widget2$getName()}}
\item \href{#method-getNodeId}{\code{Widget2$getNodeId()}}
\item \href{#method-getHtml}{\code{Widget2$getHtml()}}
\item \href{#method-getType}{\code{Widget2$getType()}}
\item \href{#method-getIoType}{\code{Widget2$getIoType()}}
\item \href{#method-isInput}{\code{Widget2$isInput()}}
\item \href{#method-isOutput}{\code{Widget2$isOutput()}}
\item \href{#method-getValue}{\code{Widget2$getValue()}}
\item \href{#method-setValue}{\code{Widget2$setValue()}}
\item \href{#method-click}{\code{Widget2$click()}}
\item \href{#method-sendKeys}{\code{Widget2$sendKeys()}}
\item \href{#method-listTabs}{\code{Widget2$listTabs()}}
\item \href{#method-uploadFile}{\code{Widget2$uploadFile()}}
\item \href{#method-clone}{\code{Widget2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new \code{Widget2}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$new(
  name,
  node_id,
  type,
  iotype = c("input", "output"),
  chromote_session
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of a Shiny widget.}

\item{\code{type}}{Widget2 type}

\item{\code{iotype}}{Input/output type.}

\item{\code{nodeId}}{Node Id of an Element from the \code{chromote} session.}

\item{\code{chromote}}{Chromote Session}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getName"></a>}}
\if{latex}{\out{\hypertarget{method-getName}{}}}
\subsection{Method \code{getName()}}{
Control id (i.e. \code{inputId} or \code{outputId} that control
was created with).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$getName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNodeId"></a>}}
\if{latex}{\out{\hypertarget{method-getNodeId}{}}}
\subsection{Method \code{getNodeId()}}{
Underlying \code{nodeId} to be used within the Chromium session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$getNodeId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHtml"></a>}}
\if{latex}{\out{\hypertarget{method-getHtml}{}}}
\subsection{Method \code{getHtml()}}{
Retrieve the underlying HTML for a widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$getHtml()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getType"></a>}}
\if{latex}{\out{\hypertarget{method-getType}{}}}
\subsection{Method \code{getType()}}{
Widget2 type, e.g. \code{textInput}, \code{selectInput}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$getType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIoType"></a>}}
\if{latex}{\out{\hypertarget{method-getIoType}{}}}
\subsection{Method \code{getIoType()}}{
Is this an input or output control?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$getIoType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isInput"></a>}}
\if{latex}{\out{\hypertarget{method-isInput}{}}}
\subsection{Method \code{isInput()}}{
Is this an input control?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$isInput()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isOutput"></a>}}
\if{latex}{\out{\hypertarget{method-isOutput}{}}}
\subsection{Method \code{isOutput()}}{
Is this an output control?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$isOutput()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getValue"></a>}}
\if{latex}{\out{\hypertarget{method-getValue}{}}}
\subsection{Method \code{getValue()}}{
Get current value of control.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$getValue()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setValue"></a>}}
\if{latex}{\out{\hypertarget{method-setValue}{}}}
\subsection{Method \code{setValue()}}{
Set value of control.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$setValue(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value to set for the widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-click"></a>}}
\if{latex}{\out{\hypertarget{method-click}{}}}
\subsection{Method \code{click()}}{
Sends a jQuery click to the element.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$click()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendKeys"></a>}}
\if{latex}{\out{\hypertarget{method-sendKeys}{}}}
\subsection{Method \code{sendKeys()}}{
Send specified key presses to control.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$sendKeys(keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{Keys to send to the widget or the app.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listTabs"></a>}}
\if{latex}{\out{\hypertarget{method-listTabs}{}}}
\subsection{Method \code{listTabs()}}{
Lists the tab names of a \code{\link[shiny:tabsetPanel]{shiny::tabsetPanel()}}.
It fails for other types of widgets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$listTabs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-uploadFile"></a>}}
\if{latex}{\out{\hypertarget{method-uploadFile}{}}}
\subsection{Method \code{uploadFile()}}{
Upload a file to a \code{\link[shiny:fileInput]{shiny::fileInput()}}.
It fails for other types of widgets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$uploadFile(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Path to file to upload}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Widget2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
